requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
backports.functools-lru-cache==2.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debacl==1.1
debugpy==1.8.16
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.3.0
executing==2.2.1
fastcore==1.8.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.55.4
fsspec==2025.5.1
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hdbscan==0.8.40
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.1
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.3.0
parso==0.8.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.4.0
prettytable==3.12.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pydantic==1.10.22
Pygments==2.19.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
subprocess32==3.5.4
sympy==1.14.0
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.16.0
threadpoolctl==3.5.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.2
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
utils==1.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wget==3.2
wrapt==1.14.2
zipp==3.21.0
zstandard==0.23.0
